+++ src/App.jsx
@@ -1,20 +1,13 @@
 import './App.css';
+import TodoList from './TodoList';
+import TodoForm from './TodoForm';
 
 function App() {
-  const todos = [
-    { id: 1, title: 'Make breakfast' },
-    { id: 2, title: 'Do laundry' },
-    { id: 3, title: 'Clean floors' },
-  ];
-
   return (
     <div>
       <h1>To-do list</h1>
-      <ul>
-        {todos.map((todo) => (
-          <li key={todo.id}>{todo.title}</li>
-        ))}
-      </ul>
+      <TodoForm />
+      <TodoList />
     </div>
   );
 }
